refactor: regenerate Rust Ratatui renderer
